bug 1063830 - fix and add pre_social_login tests
